TestOnly... methods are in their own file now
I am open to suggestions about a better file name here.
